pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "media-search"
version = "0.1.0"
description = "Media search service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlmodel>=0.0.14",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "PyYAML>=6.0",
    "pydantic-settings>=2.0",
    "pytest>=9.0.2",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "jinja2>=3.1",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-order>=1.0",
    "testcontainers[postgres]>=4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "migration: run Alembic migration tests (use: pytest -m migration)",
]
addopts = ["-m", "not migration"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
